Fix IAN and OAN side effects that clobber delayStatus
IAN and OAN were setting activeChannel->delayStatus = 0
before assigning activeChannel to the channel the instruction
was called for.

As a result of this I ran into trouble trying to do timing for
mt362x.c because DSD was clobbering delayStatus for channel 11
while doing an OAN to channel 10 to update the display.